requirements.txt
torch>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
faiss-cpu>=1.7.0
rank-bm25>=0.2.2
sentencepiece>=0.1.99
huggingface_hub>=0.16.0
gdown>=4.7.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0